added a new model to handle file explorer viewers
