+++ src/api/models/file-viewer.ts
@@ -1,0 +1,7 @@
+
+export interface IFileState {
+    blob?: Blob;
+    blobName: string;
+    blobText: any;
+    fileType: string;
+  }
