+++ src/store/explorer/actions.ts
@@ -7,6 +7,9 @@ import UITreeNodeModel, { IUITreeNode } from "../../api/models/file-explorer";
 
 // Description: Parse the files array into a File tree obj
 export const setExplorerRequest = (files: IFeedFile[], selected: IPluginItem) => action(ExplorerActionTypes.SET_EXPLORER_REQUEST, new UITreeNodeModel(files, selected).getTree() );
-// export const setExplorerSuccess = (items: IFeedFile[], selected: IPluginItem) => action( ExplorerActionTypes.SET_EXPLORER_SUCCESS, new UITreeNodeModel(items, selected).getTree());
 
+// Description: Stores the current selected file
 export const setSelectedNode = (node: IUITreeNode) => action(ExplorerActionTypes.SET_SELECTED_NODE, node);
+
+// Description: Stores the current selected folder or parent folder of the selected file
+export const setSelectedFolder = (node: IUITreeNode) => action(ExplorerActionTypes.SET_SELECTED_FOLDER, node);
