+++ src/store/plugin/saga.ts
@@ -74,7 +74,6 @@ function* handleGetPluginFiles(action: any) {
       console.error(res.error); // working user messaging
     } else {
       yield put(getPluginFilesSuccess(res));
-    //   yield put(setExplorerSuccess(res.data.results, selected)); // Structure the files for explorer
     }
   } catch (error) {
     console.error(error); // working user messaging
